requirements.txt
fpdf2==2.5.5
requests==2.31.0
Pillow==10.0.1
